86bzxxmxk Move DTOs into `contact-method` folder
