+++ src/modules/contacts/dtos/contact-method/contact-method-email.dto.ts
