+++ src/modules/contacts/dtos/contact-method/contact-method.abstract.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
-import { ContactMethodType, ContactMethodTypeEnum } from '../enums/contact-type.enum';
+import { ContactMethodType, ContactMethodTypeEnum } from '../../enums/contact-type.enum';
 import { IsEnum, IsOptional, IsString } from 'class-validator';
 
 export abstract class ContactMethodDto {
